Create basic detail screen UI
