+++ PhotoGallery/PhotoGallery/AppDelegate.swift
@@ -10,7 +10,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window = UIWindow(frame: UIScreen.main.bounds)
         window?.makeKeyAndVisible()
         let navigationController = UINavigationController(rootViewController: MainScreenViewController())
-        window?.backgroundColor = .green
         window?.rootViewController = navigationController
         
         return true
